Revert proxy and error changes
The proxy shouldn't be a static on WebAuthenticator, `OAuth1` is for
anyone to use, not just `OAuth1Authenticator`. We should probably take
this as a param in CreateRequest, to be reviewed later.

There is already an error mechanism on Authenticator, I'm not sure there's
any reason to add a separate system just for page load failures.