Add ClearCookiesBeforeLogin and ClearCookies()
The general assumption is that the user is showing the login UI to login a
new user. With that, automatically clearing the cookies before showing the
UI works to ensure they're not getting logged back into the user they're
already logged in as when switching accounts.

However, in situations where XA is only half of the story (like Azure
Mobile Services), it's useful to just keep the cookies but have a way to
clear them manually. So, in order to maintain current behavior,
ClearCookiesBeforeLogin was added and defaulted to true.