[Android] Prompt users for SSL failures
We'll go ahead and store acceptance per-certificate for the length of the
browser session. When images are linked to from HTTPS there can be a
significant number of errors, so we want to avoid showing 10 dialogs for
the same certificate failure.

The issue reporting this includes an explanation of the failure, but
Android's built-in browser does not, so we'll follow the Android
example.

Fixes #34