Removed URL-escaping of scopes parameter
URL-escaping the scopes parameter of an OAuth request to Instagram (and
possibly other services as well) causes a failure. If the scopes
parmeter is “basic+likes+comments+relationships”, it is URL-escaped to
"basic%2Blikes%2Bcomments%2Brelationships”. The auth endppint at
Instagram does not like this and replies with {"code": 400,
"error_type": "OAuthException", "error_message": "Invalid scope
field(s): basic+likes+”comments+relationships}. I imagine that other
services work similarly and also barf if the scopes param is
URL-encoded. So, I have added a flag to the OAuth2Authenticator
constructrs to enable/disable the URL-escaping of the scopes parameter.