bug fix. Prevent error dialog from beeing displayed.
When using Oauth2 with Google, we have to register a http://localhost
redirect callback, but as we’re not running a server, we can’t resolve
this url, and thus we should intercept this redirect in the WebView.
If we don’t do it, a “Cannot connect to Host” error is displayed in the
LoadFailed delegate. We could simply ignore this error, but I figured
that it would be better to just avoid loading the page. However, the
WebView will still invoke the LoadFailed delegate, but with a 102 error
code, meaning that the frame load was interrupted.
Note that even though an error message was displayed, the auth
component was still able to complete the authentication.